pyproject.toml
[project]
name = "code_annotation"
version = "1.0.0"
description = "A Python library for annotating and formatting Python code."
authors = [{name = "Leonardo Mieschendahl", email = "miesschendahl.leonardo@gmail.com"}]
dependencies = ["easy_prompting @ git+https://github.com/TheoreticalCoder/EasyPrompting.git"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"